requirements.txt
python-telegram-bot==21.10
google-generativeai==0.8.4
flask==3.1.0
sqlalchemy==2.0.37
python-dotenv==1.0.1
psycopg2-binary==2.9.10
requests==2.32.3
gunicorn==23.0.0
aiosqlite==0.20.0
